tests/wpt/web-platform-tests/css/css-pseudo/highlight-pseudos-computed.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Pseudo-Elements Test: highlight selectors getComputedStyle</title>
<link rel="help" href="https://drafts.csswg.org/css-pseudo/#highlight-selectors">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #target::selection,
  #target::target-text,
  #target::spelling-error,
  #target::grammar-error {
    background-color: green;
    color: lime;
  }
</style>
<div id="target"></div>
<script>
  for (const pseudo of ["::selection", "::target-text", "::spelling-error", "::grammar-error"]) {
    test(() => {
      let style = getComputedStyle(target, pseudo);
      assert_equals(style.backgroundColor, "rgb(0, 128, 0)", "Background color is green.");
      assert_equals(style.color, "rgb(0, 255, 0)", "Color is lime.");
    }, `getComputedStyle() for ${pseudo}`);
  }
</script>
